Refactor tray icon and hotkey logic
